Updates the tests to use the new app_context
